+++ karma.conf.js
@@ -3,58 +3,78 @@
 // For all available config options and default values, see:
 // https://github.com/karma-runner/karma/blob/stable/lib/config.js#L54
 
+module.exports = function (config) {
+  'use strict';
 
-// base path, that will be used to resolve files and exclude
-basePath = '';
+  config.set({
+    // base path, that will be used to resolve files and exclude
+    basePath: '',
 
-// list of files / patterns to load in the browser
-files = [
-  // loaded without require
-  'bower_components/es5-shim/es5-shim.js',
-  'bower_components/es5-shim/es5-sham.js',
+    frameworks: [
+      'jasmine'
+    ],
 
-  // frameworks
-  JASMINE,
-  JASMINE_ADAPTER,
-  REQUIRE,
-  REQUIRE_ADAPTER,
+    // list of files / patterns to load in the browser
+    files: [
+      // loaded without require
+      'bower_components/es5-shim/es5-shim.js',
+      'bower_components/es5-shim/es5-sham.js',
+      'bower_components/jquery/dist/jquery.js',
+      'build/flight.js',
 
-  // loaded without require
-  'bower_components/jquery/jquery.js',
-  'build/flight.js',
+      // hack to load RequireJS after the shim libs
+      'node_modules/karma-requirejs/lib/require.js',
+      'node_modules/karma-requirejs/lib/adapter.js',
 
-  // loaded with require
-  {pattern: 'lib/**/*.js', included: false},
-  {pattern: 'test/spec/**/*_spec.js', included: false},
+      // loaded with require
+      {pattern: 'lib/**/*.js', included: false},
+      {pattern: 'test/spec/**/*_spec.js', included: false},
 
-  'test/test-main.js'
-];
+      'test/test-main.js'
+    ],
 
-// use dots reporter, as travis terminal does not support escaping sequences
-// possible values: 'dots', 'progress', 'junit', 'teamcity'
-// CLI --reporters progress
-reporters = [
-  'dots'
-];
+    // list of files to exclude
+    exclude: [],
 
-// enable / disable watching file and executing tests whenever any file changes
-// CLI --auto-watch --no-auto-watch
-autoWatch = true;
+    // use dots reporter, as travis terminal does not support escaping sequences
+    // possible values: 'dots', 'progress'
+    // CLI --reporters progress
+    reporters: ['dots'],
 
-// start these browsers, currently available:
-// - Chrome
-// - ChromeCanary
-// - Firefox
-// - Opera
-// - Safari (only Mac)
-// - PhantomJS
-// - IE (only Windows)
-// CLI --browsers Chrome,Firefox,Safari
-browsers = [
-  'Chrome',
-  'Firefox'
-];
+    // enable / disable watching file and executing tests whenever any file changes
+    // CLI --auto-watch --no-auto-watch
+    autoWatch: true,
 
-// auto run tests on start (when browsers are captured) and exit
-// CLI --single-run --no-single-run
-singleRun = false;
+    // Start these browsers, currently available:
+    // - Chrome
+    // - ChromeCanary
+    // - Firefox
+    // - Opera
+    // - Safari (only Mac)
+    // - PhantomJS
+    // - IE (only Windows)
+    // CLI --browsers Chrome, Firefox, Safari
+    browsers: [
+      'Chrome',
+      'Firefox'
+    ],
+
+    // If browser does not capture in given timeout [ms], kill it
+    // CLI --capture-timeout 5000
+    captureTimeout: 20000,
+
+    // Auto run tests on start (when browsers are captured) and exit
+    // CLI --single-run --no-single-run
+    singleRun: false,
+
+    plugins: [
+      'karma-jasmine',
+      'karma-requirejs',
+      'karma-chrome-launcher',
+      'karma-firefox-launcher',
+      'karma-ie-launcher',
+      'karma-phantomjs-launcher',
+      'karma-safari-launcher'
+    ]
+  });
+};